+++ core/src/main/java/matrixstudio/formula/Reference.java
@@ -18,8 +18,8 @@ public class Reference implements Formula {
     }
 
     @Override
-    public long evaluate(Map<String, Long> context) throws EvaluationException {
-        Long result = context.get(name);
+    public int evaluate(Map<String, Integer> context) throws EvaluationException {
+        Integer result = context.get(name);
         if (result == null) throw new EvaluationException("Parameter '" +  name + "' isn't defined");
         return result;
     }
